+++ src/ui_interface.h
@@ -109,6 +109,12 @@ public:
     boost::signals2::signal<void (void)> BannedListChanged;
 };
 
+/** Show warning message **/
+void InitWarning(const std::string& str);
+
+/** Show error message **/
+bool InitError(const std::string& str);
+
 std::string AmountHighWarn(const std::string& optname);
 
 extern CClientUIInterface uiInterface;
