+++ src/wallet/wallet.h
@@ -116,12 +116,6 @@ struct CompactTallyItem
     }
 };
 
-/** Show warning message **/
-void InitWarning(const std::string& str);
-
-/** Show error message **/
-bool InitError(const std::string& str);
-
 /** A key pool entry */
 class CKeyPool
 {
